+++ pyxem/utils/indexation_utils.py
@@ -32,8 +32,8 @@ import psutil
 import scipy
 
 from pyxem.utils.expt_utils import _cart2polar
-from pyxem.utils.vector_utils import get_rotation_matrix_between_vectors
-from pyxem.utils.vector_utils import get_angle_cartesian
+from pyxem.utils.vectors import get_rotation_matrix_between_vectors
+from pyxem.utils.vectors import get_angle_cartesian
 from pyxem.utils.cuda_utils import (
     is_cupy_array,
     get_array_module,
